% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy-list.R
\name{lazy_list}
\alias{lazy_list}
\title{A lazy list}
\usage{
lazy_list(...)
}
\arguments{
\item{...}{a list of expressions that are captured and
only evaluated on demand.}
}
\description{
A lazy list is a list of rlang quosures.
Each element is only evaluated when needed. I.e. when
casted to a list (e.g. using \code{as.list}) or accessed using \code{[[}.
}
\examples{
library(lazyseq)
val <- 10
ll <- lazy_list(print(1), print(!!val), stop("will never execute"))
ll[[1]]
ll[[2]]
as.list(ll[-3])

}
